fix: reconnection strategy for comms
